--- Example/Tests/Tests-Prefix.pch
@@ -1,11 +1,0 @@
-//
-//  Prefix header
-//
-//  The contents of this file are implicitly included at the beginning of every test case source file.
-//
-
-#ifdef __OBJC__
-
-  
-
-#endif
