--- Example/Tests/Tests.m
@@ -1,9 +1,0 @@
-//
-//  KeychainSwiftAPITests.m
-//  KeychainSwiftAPITests
-//
-//  Created by Denis Krivitski on 12/02/2014.
-//  Copyright (c) 2014 Denis Krivitski. All rights reserved.
-//
-
-${TEST_EXAMPLE}
